requirements.txt
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comtypes==1.4.11
idna==3.10
Levenshtein==0.27.1
playsound==1.3.0
PyAudio==0.2.14
pycparser==2.22
pypiwin32==223
python-Levenshtein==0.27.1
pyttsx3==2.99
pywin32==311
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.4
srt==3.5.3
tqdm==4.67.1
urllib3==2.5.0
vosk==0.3.45
websockets==15.0.1
